Finished basics of Person constructor. Don't think it's working...

+++ pkg/Calculations.java
@@ -2,19 +2,19 @@ package pkg;
 
 public class Calculations {
     // Constants of conversion
-    private final static float GRAVITY = 9.80665f;
-    private final static float LBS_TO_KG = 0.4535924f;
-    private final static float IN_TO_CM = 2.54f;
+    private final static double GRAVITY = 9.80665f;
+    private final static double LBS_TO_KG = 0.4535924f;
+    private final static double IN_TO_CM = 2.54f;
     
-    public static float convertToCm(float inches) {
+    public static double convertToCm(double inches) {
 		return (inches * IN_TO_CM);
 	}
 
-    public static float convertToKg(float lbs) {
+    public static double convertToKg(double lbs) {
 		return (lbs * LBS_TO_KG);
 	}
 
-	public static float getMoment(float length, float distance) {
+	public static double getMoment(double length, double distance) {
 		return (length * distance);
 	}
